+++ src/routers/AppRouter.tsx
@@ -30,10 +30,9 @@ import DBTIEditPage from "@/pages/profile/edit/DBTI/DBTIEditPage";
 import ProfileEditPage from "@/pages/profile/edit/DBTI/ProfileEditPage";
 // import DBTIResultPage from "@/pages/profile/edit/DBTI/DBTIResultPage";
 import DBTIPage from "@/pages/profile/DBTI/DBTIPage";
-import ManageRoomMatched from "@/pages/manageroom/ManageRoomMatched";
-import ManageRoomDefault from "@/pages/manageroom/ManageRoomDefault";
+import ManageRoomMatched from "@/pages/manage/ManageRoomMatched";
+import ManageRoomDefault from "@/pages/manage/ManageRoomDefault";
 import Room from "@/pages/room/Room";
-import DashboardPage from "@/pages/manage/dashboard/DashboardPage";
 
 export const router = createBrowserRouter([
   {
@@ -197,12 +196,6 @@ export const router = createBrowserRouter([
         path: "manage/default",
         element: (
           <ManageRoomDefault />
-        ),
-      },
-      {
-        path: "manage/dashboard",
-        element: (
-          <DashboardPage />
         ),
       },
       {
